+++ src/foundations/index.ts
@@ -1,3 +1,4 @@
+export { default as colors } from './colors'
 export { default as fonts } from './fonts'
 export { default as sizes } from './sizes'
 export { default as space } from './space'